requirements.txt
python-switchbot==2.3.1
